+++ js/game/game-data.js
@@ -1,0 +1,13 @@
+export const INITIAL_STATE = {
+  mistakes: 0,
+  time: 300,
+  score: 0,
+  level: `artist-1`,
+};
+
+export const GAME_RULE = Object.freeze({
+  AVAILABLE_MISTAKES: 3,
+  FAST_TIME: 30,
+  MIN_ANSWERS: 10,
+});
+
