+++ js/game/game-presenter.js
@@ -29,8 +29,6 @@ class GamePresenter {
       case `genre`:
         game = new GenreView(this.model.state, this.model.currentLevelData);
         break;
-      default:
-        throw new Error(`Wrong game level type`);
     }
     return game;
   }
@@ -123,7 +121,9 @@ class GamePresenter {
       isCorrect: isCorrectAnswer,
       time: this.levelTime,
     });
+
     this.model.changeLevel();
+
     if (this.model.state.level !== `result`) {
       this.stopLevelTimer();
       this.nextGame();
@@ -131,8 +131,8 @@ class GamePresenter {
       this.stopGame();
       this.isOver(this.model.state, this.model.userAnswers);
     }
-
   }
+
   isOver() {}
 }
