--- js/game/game.js
@@ -1,12 +1,0 @@
-export const INITIAL_STATE = Object.freeze({
-  mistakes: 0,
-  time: 300,
-  score: 0,
-  level: `artist-1`,
-});
-
-export const GAME_RULE = {
-  AVAILABLE_MISTAKES: 3,
-  FAST_TIME: 30
-};
-
