+++ js/views/abstract-view.js
@@ -1,4 +1,9 @@
-import {getElementFromTemplate} from '../service/template';
+const getElementFromTemplate = (screenTemplate) => {
+  const newScreen = document.createElement(`section`);
+  newScreen.className = `screen`;
+  newScreen.innerHTML = screenTemplate;
+  return newScreen;
+};
 
 class AbstractView {
   constructor() {
